### 2017-03-09 c788f5f
Fix PyPlot for 0.6
