### 2017-03-10 868efa8
Use compat for abstract types
