### 2017-03-11 b0d2ccd
Cleaner world-age fix for pyplot
