### 2017-03-12 #204 @whedon
[REVIEW]: Weave.jl: Scientific Reports Using Julia
**Submitting author:** @mpastell (<a href="http://orcid.org/0000-0002-5810-4801">Matti Pastell</a>)
**Repository:** <a href="https://github.com/mpastell/weave.jl" target ="_blank">https://github.com/mpastell/weave.jl</a>
**Version:** v0.4.1
