### 2017-03-13 888b732
Fix output for `term=true`. Closes #82.
