### 2017-03-13 0ffa04b
Fix tests
