+++ test/formatter_test.jl
@@ -24,13 +24,13 @@ f = Weave.format_chunk(dchunk, docformat.formatdict, docformat)
 parsed = Weave.read_doc("documents/chunk_options.noweb")
 doc = Weave.run(parsed, doctype = "md2html")
 
-c_check = "<pre class='hljl'>\n<span class='hljl-nf'>println</span><span class='hljl-p'>(</span><span class='hljl-n'>x</span><span class='hljl-p'>)</span>\n</pre>\n"
+c_check = "<pre class='hljl'>\n<span class='hljl-n'>x</span><span class='hljl-t'> </span><span class='hljl-oB'>=</span><span class='hljl-t'> </span><span class='hljl-p'>[</span><span class='hljl-ni'>12</span><span class='hljl-p'>,</span><span class='hljl-t'> </span><span class='hljl-ni'>10</span><span class='hljl-p'>]</span><span class='hljl-t'>\n</span><span class='hljl-nf'>println</span><span class='hljl-p'>(</span><span class='hljl-n'>y</span><span class='hljl-p'>)</span>\n</pre>\n"
 doc.format.formatdict[:theme] = doc.highlight_theme
 c = Weave.format_code(doc.chunks[3].content, doc.format)
 @test c_check == c
 
-o_check = "\nprintln&#40;x&#41;"
-o = Weave.format_output(doc.chunks[3].content, doc.format)
+o_check = "\nprintln&#40;x&#41;\n"
+o = Weave.format_output(doc.chunks[4].content, doc.format)
 @test o_check == o
 
 doc.template = "templates/mini.tpl"
@@ -43,17 +43,16 @@ doc = Weave.run(parsed, doctype = "md2tex")
 
 c_check = "\\begin{lstlisting}\n(*@\\HLJLnf{println}@*)(*@\\HLJLp{(}@*)(*@\\HLJLn{x}@*)(*@\\HLJLp{)}@*)\n\\end{lstlisting}\n"
 doc.format.formatdict[:theme] = doc.highlight_theme
-c = Weave.format_code(doc.chunks[3].content, doc.format)
+c = Weave.format_code(doc.chunks[4].content, doc.format)
 @test c_check == c
 
-o_check = "\nx = [12, 10]\nprintln(y)"
-o = Weave.format_output(doc.chunks[2].content, doc.format)
+o_check = "\nx = [12, 10]\nprintln(y)\n"
+o = Weave.format_output(doc.chunks[3].content, doc.format)
 @test o_check == o
 
 doc.template = "templates/mini.tpl"
 rendered = Weave.render_doc("Hello", doc, doc.format)
 @test rendered == "\nHello\n"
-
 
 # Test header parsing and stripping
 header = """
