+++ test/documents/gadfly_formats_test_mmd_ref.md
@@ -19,9 +19,9 @@ plot(x=x, y = sin(x), Geom.line)
 
 ````julia
 julia> x = linspace(0, 2π, 200)
-
 200-element LinSpace{Float64}:
  0.0,0.0315738,0.0631476,0.0947214,0.126295,…,6.18846,6.22004,6.25161,6.28319
+
 julia> plot(x=x, y = sin(x), Geom.line)
 
 ````
@@ -33,9 +33,10 @@ julia> plot(x=x, y = sin(x), Geom.line)
 
 ````julia
 julia> y = 20
-
 20
+
 julia> plot(x=x, y = cos(x), Geom.line)
+
 ````
 
 
