+++ test/documents/gadfly_formats_test_ref.rst
@@ -3,7 +3,7 @@
     using Gadfly
     x = linspace(0, 2π, 200)
     plot(x=x, y = sin(x), Geom.line)
-
+    
 
 
 .. figure:: figures/gadfly_formats_test_sin_fun_1.png
@@ -25,9 +25,9 @@
 .. code-block:: julia
 
 julia> x = linspace(0, 2π, 200)
-
 200-element LinSpace{Float64}:
  0.0,0.0315738,0.0631476,0.0947214,0.126295,…,6.18846,6.22004,6.25161,6.28319
+
 julia> plot(x=x, y = sin(x), Geom.line)
 
 
@@ -40,9 +40,10 @@ julia> plot(x=x, y = sin(x), Geom.line)
 .. code-block:: julia
 
 julia> y = 20
-
 20
+
 julia> plot(x=x, y = cos(x), Geom.line)
+
 
 
 
@@ -54,7 +55,7 @@ julia> plot(x=x, y = cos(x), Geom.line)
     
     x = linspace(0, 2π, 200)
     plot(x=x, y = sin(x), Geom.line)
-
+    
 
 
 .. image:: figures/gadfly_formats_test_5_1.png
@@ -65,7 +66,7 @@ julia> plot(x=x, y = cos(x), Geom.line)
     
     y = 20
     plot(x=x, y = cos(x), Geom.line)
-
+    
 
 
 .. image:: figures/gadfly_formats_test_5_2.png

[…]
