### 2017-03-13 mpastell/Weave.jl#85 @vchuravy
Review for JOSS
I just started my review for JOSS (https://github.com/openjournals/joss-reviews/issues/204) and I will note down issues as I encounter them:

- [x] I get warnings for Base methods that are overwritten on both 0.5 and 0.6? Is this necessary? (`latex` and `latexinline`)
- [x] The copyright notice in LICENSE.md should be updated from `Copyright (c) 2014: Matti Pastell.` to `Copyright (c) 2014-2017: Matti Pastell and other contributors.`
- [x] There are a bunch of deprecation warnings for Julia v0.6 and it would be great if these could be addressed.
- [x] The optional dependencies for the examples in https://github.com/mpastell/Weave.jl#usage e.g. `Gadlfy`, `Cairo`, `Fontconfig` should be mentioned.
- [x] I also get the following warning on v0.5 that should be fixed.
```
WARNING: eval from module Base to Weave:
Expr(:->, :x, Expr(:block, Expr(:line, 554, :/home/wallnuss/.julia/v0.5/Documenter/src/DocChecks.jl)::Any, Expr(:block, :_ = :have_color, Expr(:global, :have_color = :x)::Any, :_)::Any)::Any)::Any
  ** incremental compilation may be broken for this module **
```
