### 2017-03-13 mpastell/Weave.jl#85 @mpastell
Hi,
Thanks for being so quick to start and for the remarks. A lot of the issues are already fixed in current Weave master. Here are few first responses:

1. Base.markdown methods need to be overwritten for 0.5, but not 0.6 (fixed in https://github.com/JuliaLang/julia/pull/19842 and https://github.com/JuliaLang/julia/pull/19832). I don't get any warnings with version 0.6.0-pre.alpha.34.
2. I've started fixing deprecation warnings in current master. I think all of the warnings from Weave are fixed, but most of the depencies haven't been updated (or at least not tagged yet). I will try to get a "warning free" version released when Julia 0.6 is released.
3. The warning about precompilation has appeared after last release and has been fixed current in master by setting `__precompile__(false)`.
