+++ LICENSE.md
@@ -1,6 +1,6 @@
 The Weave.jl package is licensed under the MIT "Expat" License:
 
-> Copyright (c) 2014: Matti Pastell.
+> Copyright (c) 2014-2017: Matti Pastell.
 >
 > Permission is hereby granted, free of charge, to any person obtaining
 > a copy of this software and associated documentation files (the
