### 2017-03-13 mpastell/Weave.jl#85 @mpastell
I just updated the license as well.
