### 2017-03-13 09ecdb6
Disable plots tests
