### 2017-03-13 b2ffefd
Remove Winston support, closes #58
