+++ test/runtests.jl
@@ -1,12 +1,6 @@
 using Weave
 using Base.Test
 
-# Running Changin plotlib in tests segfault, unless they are run
-# using separate processes.
-#run(`julia --code-coverage=user -e 'include("winston_formats.jl")'`)
-#run(`julia --code-coverage=user -e 'include("pyplot_formats.jl")'`)
-
-
 info("Test: Chunk options")
 include("chunk_options.jl")
 
@@ -34,9 +28,5 @@ if VERSION < v"0.6-"
 
   #info("Test: Weaving with Plots.jl")
   #include("plotsjl_test.jl")
-  include("publish_test.jl")
+  #include("publish_test.jl")
 end
-
-#Fails on travis, works locally.
-#info("Test: Weaving with Winston")#
-#include("winston_formats.jl")
