### 2017-03-13 d729d4f
Add 0.6 to Travis
