+++ .travis.yml
@@ -2,6 +2,7 @@
 language: julia
 julia:
   - 0.5
+  - 0.6
   - nightly
 matrix:
   allow_failures:
