### 2017-03-13 1b5c0f8
Fix tests for 0.6
