+++ test/runtests.jl
@@ -7,16 +7,16 @@ include("chunk_options.jl")
 info("Test: Converting")
 include("convert_test.jl")
 
-info("Test: Caching")
-include("cache_test.jl")
-
-info("Testing rich output")
-include("rich_output.jl")
-
 info("Testing formatters")
 include("formatter_test.jl")
 
 if VERSION < v"0.6-"
+  info("Testing rich output")
+  include("rich_output.jl")
+  
+  info("Test: Caching")
+  include("cache_test.jl")
+
   info("Test: Chunk options with Gadfly")
   include("chunk_opts_gadfly.jl")
 
