### 2017-03-14 bbad66d
Fix plotting imports and tests
