+++ src/gadfly.jl
@@ -1,10 +1,10 @@
-using Gadfly
+import Gadfly
 
 Gadfly.set_default_plot_format(:svg)
 
 #Gadfly doesn't call the default display methods, this catches
 #all Gadfly plots
-function Base.display(report::Report, m::MIME"image/svg+xml", p::Plot)
+function Base.display(report::Report, m::MIME"image/svg+xml", p::Gadfly.Plot)
 
     chunk = report.cur_chunk
     full_name, rel_name = get_figname(report, chunk)
@@ -14,28 +14,28 @@ function Base.display(report::Report, m::MIME"image/svg+xml", p::Plot)
 
     report.fignum += 1
 
-    w = chunk.options[:fig_width]inch
-    h = chunk.options[:fig_height]inch
+    w = chunk.options[:fig_width]Gadfly.inch
+    h = chunk.options[:fig_height]Gadfly.inch
     format = chunk.options[:fig_ext]
     dpi = chunk.options[:dpi]
 
     #This is probably not the correct way to handle different formats, but it works.
     if format == ".png"
         try
-            draw(PNG(full_name, w, h, dpi=dpi), p)
+            Gadfly.draw(Gadfly.PNG(full_name, w, h, dpi=dpi), p)
         catch
-            draw(PNG(full_name, w, h), p) #Compose < 0.3.1, Gadfly < 0.3.1
+            Gadfly.draw(Gadfly.PNG(full_name, w, h), p) #Compose < 0.3.1, Gadfly < 0.3.1
         end
     elseif format == ".pdf"
-        draw(PDF(full_name, w, h), p)
+        Gadfly.draw(Gadfly.PDF(full_name, w, h), p)
     elseif format == ".ps"
-        draw(PS(full_name, w, h), p)
+        Gadfly.draw(Gadfly.PS(full_name, w, h), p)
     elseif format == ".svg"
-        draw(SVG(full_name, w, h), p)
+        Gadfly.draw(Gadfly.SVG(full_name, w, h), p)
     elseif format == ".js.svg"
-        draw(SVGJS(full_name, w, h), p)
+        Gadfly.draw(Gadfly.SVGJS(full_name, w, h), p)
     elseif format == ".tex"
-        draw(PGF(full_name, w, h, true ), p)
+        Gadfly.draw(Gadfly.PGF(full_name, w, h, true ), p)
     else
         warn("Can't save figure. Unsupported format")
     end
