+++ src/pyplot.jl
@@ -1,4 +1,4 @@
-using PyPlot
+import PyPlot
 
 function savefigs_pyplot(report::Report)
     chunk = report.cur_chunk
@@ -9,12 +9,12 @@ function savefigs_pyplot(report::Report)
     chunkid = (chunk.options[:name] == nothing) ? chunk.number : chunk.options[:name]
     #Iterate over all open figures, save them and store names
 
-    for fig = plt[:get_fignums]()
+    for fig = PyPlot.plt[:get_fignums]()
         full_name, rel_name = get_figname(report, chunk, fignum=fig)
-        savefig(full_name, dpi=chunk.options[:dpi])
+        PyPlot.savefig(full_name, dpi=chunk.options[:dpi])
         push!(report.figures, rel_name)
         report.fignum += 1
-        plt[:draw]()
-        plt[:close]()
+        PyPlot.plt[:draw]()
+        PyPlot.plt[:close]()
     end
 end
