### 2017-03-14 mpastell/Weave.jl#85 @vchuravy
Ok that sounds good once I am done with my review it would be great if you could release a new version.
