### 2017-03-14 #204 @vchuravy
@arfon How would I update the version number of the package for submission? I am nearly done with my review, but I think a point or minor release would be good to give a smooth user experience.
