### 2017-03-14 #204 @vchuravy
Also didn't we used to generate a preview of the PDF?
