### 2017-03-14 #204 @arfon
> @arfon How would I update the version number of the package for submission? I am nearly done with my review, but I think a point or minor release would be good to give a smooth user experience.

@vchuravy - I can do this. What's the new release number?

> Also didn't we used to generate a preview of the PDF?

Yeah, I usually do this at the end of the review during processing now. We plan to automate this in the future.
