### 2017-03-14 38d3b04
Only strip header from md2* formats, fix pandoc2html
