### 2017-03-14 1a54d6a
Add citeproc filter to pandoc, fixes #50
