### 2017-03-14 5513656
Add info about depencies #85
