### 2017-03-14 mpastell/Weave.jl#85 @mpastell
I have now added the info about depencies for examples to readme and docs. Are there any further issues?

I will tag a new version v0.4.1 after the review is finished.
