### 2017-03-14 c179d30
Update news and bump docs version
