### 2017-03-14 1915501
More tests for 0.6
