+++ test/rich_output.jl
@@ -2,10 +2,12 @@ using Weave
 using Base.Test
 
 function mmtest(source, resfile, doctype)
+  VER = "$(VERSION.major).$(VERSION.minor)"
   weave("documents/$source", out_path = "documents/multimedia/$resfile", doctype=doctype, plotlib=nothing)
   result =  readstring(open("documents/multimedia/$resfile"))
-  ref =  readstring(open("documents/multimedia/$resfile.ref"))
+  ref =  readstring(open("documents/multimedia/$VER/$resfile.ref"))
   @test result == ref
+  rm("documents/multimedia/$resfile")
 end
 
 mmtest("rich_output.jmd", "rich_output.md", "pandoc")
