### 2017-03-14 97f309a
Fix test cleaning on windows
