### 2017-03-14 mpastell/Weave.jl#85 @mpastell
Many thanks for the review! I have now tagged v0.4.1 with these corrections and few other bug fixes.
