### 2017-03-15 #204 @arfon
@vchuravy - here's the compiled paper PDF for your review [10.21105.joss.00204.pdf](https://github.com/openjournals/joss-reviews/files/844291/10.21105.joss.00204.pdf)
