### 2017-03-15 mpastell/Weave.jl#85 @vchuravy
One more thing I just noticed. The proper citation info for the Julia paper is
```
@article{doi:10.1137/141000671,
author = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
title = {Julia: A Fresh Approach to Numerical Computing},
journal = {SIAM Review},
volume = {59},
number = {1},
pages = {65-98},
year = {2017},
doi = {10.1137/141000671},
}
```

after: http://epubs.siam.org/doi/10.1137/141000671
