### 2017-03-15 0131bae
Updated citation #85
