+++ paper/paper.bib
@@ -1,11 +1,12 @@
 @article{julia,
-	Eprinttype = {arXiv},
-	Author = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
-	Eprint = {1411.1607},
-	Month = {November},
-	Eprintclass = {cs.MS},
-	Title = {{J}ulia: A Fresh Approach to Numerical Computing},
-	Year = {2014}
+	author = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
+	title = {Julia: A Fresh Approach to Numerical Computing},
+	journal = {SIAM Review},
+	volume = {59},
+	number = {1},
+	pages = {65-98},
+	year = {2017},
+	doi = {10.1137/141000671},
 }
 
 @Book{knitr,
