### 2017-03-15 mpastell/Weave.jl#85 @mpastell
Thanks for letting me know, I fixed it.
